+++ src/components/main/PopularPosts.jsx
@@ -1,7 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import supabase from 'lib/supabaseClient';
+import { useQuery } from 'react-query';
+import { getPosts } from 'api/main';
 
 const PopularPosts = () => {
+  // supabase
   const [posts, setPosts] = useState([]);
 
   useEffect(() => {
@@ -20,7 +23,19 @@ const PopularPosts = () => {
     getPosts();
   }, []);
 
-  console.log(posts);
+  // react-query
+  // const { isLoading, isError, data } = useQuery('posts', getPosts);
+
+  // if (isLoading) {
+  //   return <p>Loading…</p>;
+  // }
+
+  // if (isError) {
+  //   return <p>Error</p>;
+  // }
+
+  // const posts = data.data;
+
   const popularPosts = posts.sort((a, b) => b.like - a.like).slice(0, 2);
 
   return (
