+++ src/components/mypage/Profile.jsx
@@ -51,7 +51,7 @@ const Profile = () => {
 
   return (
     // 테스트 코드
-    <>
+    <div>
       <p style={{ padding: '10px' }}>[TEST DATA]</p>
       {members.map((member) => (
         <div style={{ padding: '10px' }} key={member.id}>
@@ -99,7 +99,7 @@ const Profile = () => {
         </form>
       </div> */}
       <MyPost />
-    </>
+    </div>
   );
 };
 export default Profile;
