+++ src/pages/Detail.jsx
@@ -4,10 +4,10 @@ import React from 'react';
 
 const Detail = () => {
   return (
-    <>
+    <div>
       <Post />
       <Comments />
-    </>
+    </div>
   );
 };
